Скомпилировалась библиотека

+++ Tyuiu.MedvedevKA.Sprint3.Task0.V1.Lib/DataService.cs
@@ -9,7 +9,7 @@ namespace Tyuiu.MedvedevKA.Sprint3.Task0.V1.Lib
 {
     public class DataService : ISprint3Task0V1
     {
-        public double GetSumSeries(double value,int startValue,int stopValue)
+        public double GetSumSeries(int value,int startValue,int stopValue)
         {
             double sumSeries = 0;
             int i;
